+++ lib/conjure/data_set.rb
@@ -1,7 +1,0 @@
-module Conjure
-  class DataSet
-    def self.find(options = {})
-      []
-    end
-  end
-end