+++ lib/conjure/local_docker.rb
@@ -1,0 +1,14 @@
+module Conjure
+  class LocalDocker
+    def run(command)
+      `#{command}`
+    end
+
+    def with_directory(path)
+      yield path
+    end
+
+    def ip_address
+    end
+  end
+end