+++ lib/conjure/provision.rb
@@ -1,1 +1,0 @@
-require "conjure/provision/instance"